Fix sphinx warnings
Moved task_geo.data_sources.covid.fr_covidata -> task_geo.data_sources.covid.fr_covid and split in multiple files
Moved task_geo.data_sources.covid.spain -> task_geo.data_sources.covid.es_covid
Fixing issues with docstrings formatting on different modules
Add example notebooks to the documentation index
