+++ task_geo/data_sources/covid/es_covid/es_covid_connector.py
@@ -2,13 +2,14 @@ import pandas as pd
 
 
 def es_covid_connector():
-    """Retrieves data from https://covid19.isciii.es
+    """Retrieve data from https://covid19.isciii.es.
 
     Arguments:
         None
 
     Returns:
         pandas.DataFrame
+
     """
     url = "https://covid19.isciii.es/resources/serie_historica_acumulados.csv"
     return pd.read_csv(url, encoding="latin_1")
