+++ task_geo/data_sources/covid/fr_covid/__main__.py
