+++ task_geo/data_sources/covid/fr_covid/datapackage.json
