--- task_geo/data_sources/covid/fr_covidata/__init__.py
@@ -1,3 +1,0 @@
-from task_geo.data_sources.covid.fr_covidata.fr_covidata import fr_covidata
-
-__all__ = ['fr_covidata']
